+++ src/assets/index.js
@@ -1,2 +1,5 @@
 import logo from "./logo.png";
+import icon_location_map from "./icon_location_map.svg"
+
 export const LOGO = logo;
+export const ICON_LOCATION_MAP = icon_location_map;
